refactor(web): centralize CASE_TYPES, OUTCOMES, MOTION_TYPES filter arrays (#2024)
Move the three filter option arrays from SearchPage.tsx, RulingsFeed.tsx,
and CasesList.tsx into the shared filter-options.ts module. This eliminates
triple-maintenance risk when filter values change.
